+++ standingsrequests/decorators.py
@@ -1,15 +1,18 @@
-import logging
 from functools import wraps
 
 from django.utils.decorators import available_attrs
+
+from allianceauth.services.hooks import get_extension_logger
 from esi.decorators import _check_callback
 from esi.models import Token
 from esi.views import select_token, sso_redirect
 
+from . import __title__
 from .managers.standings import StandingsManager
+from .utils import LoggerAddTag
 
 
-logger = logging.getLogger(__name__)
+logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 
 
 def token_required_by_state(new=False):